+++ providers/hashicorp/docs/changelog.rst
@@ -27,6 +27,19 @@
 Changelog
 ---------
 
+4.8.2
+.....
+
+Misc
+~~~~
+
+* ``Bump the minimum apache-airflow-providers-common-compat version to 1.12.0 (#72503)``
+
+.. Below changes are excluded from the changelog. Move them to
+   appropriate section above if needed. Do not delete the lines(!):
+   * ``Revert "[main] Upgrade important CI environment (#73308)" (#73621)``
+   * ``[main] Upgrade important CI environment (#73308)``
+
 4.8.1
 .....